fix: Move `Index` to a new file
